Fix :Theora video playback(incorrect fps), add support MT; Fix : MP4Splitter - crash on some files; Support only one video stream output; More user-friendly track name; remove some warning;
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@3975 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/mplayerc/MainFrm.cpp
@@ -8764,7 +8764,7 @@ public:
 
 	// ISequentialStream
 	STDMETHODIMP Read(void* pv, ULONG cb, ULONG* pcbRead) {
-		__int64 cbRead = min(m_data.GetCount() - m_pos, (__int64)cb);
+		__int64 cbRead = min((__int64)(m_data.GetCount() - m_pos), (__int64)cb);
 		cbRead = max(cbRead, 0);
 		memcpy(pv, &m_data[(INT_PTR)m_pos], (int)cbRead);
 		if (pcbRead) {
